Refactor kmer generators Add seq-thread target
