+++ src/sglib/graph/SequenceGraph.hpp
@@ -47,7 +47,7 @@ public:
         return nodes == o.nodes;
     }
 
-    SequenceGraph(){};
+    SequenceGraph() {};
     SequenceGraph(const SequenceGraph &sg) = delete; // Avoid implicit generation of the copy constructor.
     //=== I/O functions ===
     void load_from_gfa(std::string filename);
