+++ src/sglib/mappers/LinkedReadMapper.hpp
@@ -17,63 +17,6 @@
  * Supports partial remapping of unmapped reads or of a selection list.
  */
 class LinkedReadMapper {
-    class StreamKmerFactory : public  KMerFactory {
-    public:
-        explicit StreamKmerFactory(uint8_t k) : KMerFactory(k){}
-        inline void produce_all_kmers(const char * seq, std::vector<KmerIDX> &mers){
-            // TODO: Adjust for when K is larger than what fits in uint64_t!
-            last_unknown=0;
-            fkmer=0;
-            rkmer=0;
-            auto s=seq;
-            while (*s!='\0' and *s!='\n') {
-                //fkmer: grows from the right (LSB)
-                //rkmer: grows from the left (MSB)
-                fillKBuf(*s, fkmer, rkmer, last_unknown);
-                if (last_unknown >= K) {
-                    if (fkmer <= rkmer) {
-                        // Is fwd
-                        mers.emplace_back(fkmer);
-                        mers.back().contigID=1;
-                    } else {
-                        // Is bwd
-                        mers.emplace_back(rkmer);
-                        mers.back().contigID=-1;
-                    }
-                }
-                ++s;
-            }
-        }
-    };
-
-    class StreamKmerFactory128 : public  KMerFactory128 {
-    public:
-        explicit StreamKmerFactory128(uint8_t k) : KMerFactory128(k){}
-        inline void produce_all_kmers(const char * seq, std::vector<KmerIDX128> &mers){
-            // TODO: Adjust for when K is larger than what fits in uint64_t!
-            last_unknown=0;
-            fkmer=0;
-            rkmer=0;
-            auto s=seq;
-            while (*s!='\0' and *s!='\n') {
-                //fkmer: grows from the right (LSB)
-                //rkmer: grows from the left (MSB)
-                fillKBuf(*s, fkmer, rkmer, last_unknown);
-                if (last_unknown >= K) {
-                    if (fkmer <= rkmer) {
-                        // Is fwd
-                        mers.emplace_back(fkmer);
-                        mers.back().contigID=1;
-                    } else {
-                        // Is bwd
-                        mers.emplace_back(rkmer);
-                        mers.back().contigID=-1;
-                    }
-                }
-                ++s;
-            }
-        }
-    };
 
 public:
     LinkedReadMapper(SequenceGraph &_sg, LinkedReadsDatastore &_datastore) : sg(_sg),datastore(_datastore){
